Remove cast to short to allow storing second offset correctly.

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/index/schema/ZonedDateTimeLayout.java
@@ -96,7 +96,7 @@ class ZonedDateTimeLayout extends SchemaLayout<ZonedDateTimeSchemaKey>
 
     private int asZoneOffset( int encodedZone )
     {
-        return (short) encodedZone;
+        return encodedZone;
     }
 
     private short asZoneId( int encodedZone )
